+++ test/components/Feature.spec.tsx
@@ -1,5 +1,4 @@
-import { render, cleanup } from '@testing-library/react';
-import '@testing-library/jest-dom/extend-expect';
+import { render, screen } from '@testing-library/react';
 import { faBook } from '@fortawesome/free-solid-svg-icons';
 import Feature from '../../src/components/Feature';
 
@@ -7,13 +6,11 @@ describe('<Feature />', () => {
   const theTitle = 'The title';
   const theIcon = faBook;
 
-  afterEach(cleanup);
-
   it('properly renders provided children inside a section', () => {
-    const { getByText, container } = render(<Feature icon={theIcon} title={theTitle}>This is the feature</Feature>);
+    const { container } = render(<Feature icon={theIcon} title={theTitle}>This is the feature</Feature>);
 
-    expect(getByText('This is the feature')).toBeInTheDocument();
-    expect(getByText(theTitle)).toBeInTheDocument();
+    expect(screen.getByText('This is the feature')).toBeInTheDocument();
+    expect(screen.getByText(theTitle)).toBeInTheDocument();
     expect(container.querySelector('.fa-book')).toBeInTheDocument();
   });
 });